chore: revamp structure
