+++ utils/upEncrypted.js
@@ -1,17 +1,5 @@
-const { ethers } = require("ethers");
 const lighthouse = require("@lighthouse-web3/sdk");
-
-let hashValCID;
-const sign_auth_message = async (publicKey, privateKey) => {
-  const provider = new ethers.providers.JsonRpcProvider(
-    "https://polygon-mumbai.g.alchemy.com/v2/XU1AlAfY0g0HRHrfh7ykOGqOofGupyZm"
-  );
-  const signer = new ethers.Wallet(privateKey, provider);
-  const messageRequested = (await lighthouse.getAuthMessage(publicKey)).data
-    .message;
-  const signedMessage = await signer.signMessage(messageRequested);
-  return signedMessage;
-};
+const signAuthMessage = require("./signAuthMessage")
 
 const deployEncrypted = async () => {
   const path =
@@ -20,7 +8,7 @@ const deployEncrypted = async () => {
   const publicKey = "0xf6f1058c985572422b6ed0B5Fd4Af2Ba2704922B";
   const privateKey =
     "f80f8f01a2a34a6977d09f84ecf03eadcd4a4f73a52cee71218504964c6c33dc";
-  const signed_message = await sign_auth_message(publicKey, privateKey);
+  const signed_message = await signAuthMessage(publicKey, privateKey);
 
   const response = await lighthouse.uploadEncrypted(
     path,
@@ -41,6 +29,5 @@ const deployEncrypted = async () => {
   */
 };
 
-deployEncrypted();
-
-modules.export = { hashValCID };
+// deployEncrypted();
+module.exports = deployEncrypted;
